% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baf_download.R
\name{baf_download}
\alias{baf_download}
\title{Download a file}
\usage{
baf_download(url, path, overwrite = FALSE, quiet = FALSE)
}
\arguments{
\item{url}{a URL}

\item{path}{a file path}

\item{overwrite}{should the file at path be overwritten if it already exists? Default is FALSE.}

\item{quiet}{should output and progress bar be suppressed?}
}
\value{
a path to the downloaded file, invisibly
}
\description{
Download a file
}
\examples{
tf <- tempfile(fileext = '.zip')
ex <- 'https://www2.census.gov/geo/docs/maps-data/data/baf2020/BlockAssign_ST01_AL.zip'
baf_download(url = ex, path = tf)
}
