% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baf_download.R
\name{baf_download_path}
\alias{baf_download_path}
\title{Get download location for \code{baf} files}
\usage{
baf_download_path()
}
\value{
a path
}
\description{
Get download location for \code{baf} files
}
\examples{
baf_download_path()
}
