% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean_bafs}
\alias{clean_bafs}
\title{Clean Block Assignment Files}
\usage{
clean_bafs(bafs, state = NULL)
}
\arguments{
\item{bafs}{list of block assignment files, named by type}

\item{state}{if provided, adds the first two digits of the state FIPS to each code}
}
\value{
a tibble with two columns
\itemize{
\item GEOID: block GEOIDs
\item target: output GEOIDs for the target geography
}
}
\description{
Clean Block Assignment Files
}
\examples{
baf('WA') |> clean_bafs('WA')
}
