% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baf_cd.R
\name{download_baf_cd113}
\alias{download_baf_cd113}
\title{Download 2010 Block Assignment File to 113th Congressional Districts}
\usage{
download_baf_cd113()
}
\value{
path to new zip directory, invisibly
}
\description{
Download 2010 Block Assignment File to 113th Congressional Districts
}
