% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{example_baf}
\alias{example_baf}
\title{Example BAF data from Adams County, WA}
\usage{
data('example_baf')
}
\value{
tibble with state identifying information
}
\description{
list of tibbles with 1,562 rows each
}
\examples{
data('example_baf')
}
\concept{example_baf}
