% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GriffingM1M1.R
\name{GriffingM1M1}
\alias{GriffingM1M1}
\title{Girffing, Hayman ve diğer ıslah desenlerinin analizi}
\usage{
GriffingM1M1(data)
}
\arguments{
\item{data}{A \code{data.frame}}

\item{Rep}{Replicate as factor}

\item{Cross1}{Cross 1 as factor}

\item{Cross2}{Cross 2 as factor}

\item{Method}{Method for Diallel Analysis using Griffing's approach.
It can take \strong{1}, \strong{2}, \strong{3}, or \strong{4}
 as argument depending on the method being used.
\enumerate{
\item Method-I (Parents + \eqn{F_{1}}'s + reciprocals);
\item Method-II (Parents and one set of \eqn{F_{1}}'s);
\item Method-III (One set of \eqn{F_{1}}'s and reciprocals);
\item Method-IV (One set of \eqn{F_{1}}'s only).
}}

\item{Model}{Model for Diallel Analysis using Griffing's approach.
It can take \strong{1} or \strong{2} as arguments depending on the model being used.
\enumerate{
\item Fixed  Effects Model;
\item Random Effects Model.
}}
}
\description{
Bu paket ıslah denemelerinin analizi için geliştirilmiştir.
}
\examples{

# Example1

GriffingM1M1(data)



}
