% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HaymanPlot.R
\name{HaymanPlot}
\alias{HaymanPlot}
\title{Girffing, Hayman ve diğer ıslah desenlerinin analizi}
\usage{
HaymanPlot(data.frame)
}
\arguments{
\item{VOLO}{<- op13()$VOLO}

\item{In.Value}{<- op13()$In.Value}

\item{a}{<- op13()$a}

\item{b}{<- op13()$b}

\item{Wr.Vr}{<- op()$Wr.Vr.Table}

\item{data}{A \code{data.frame}}
}
\description{
Bu paket ıslah denemelerinin analizi için geliştirilmiştir.
}
\examples{

# Example1

HaymanPlot(data.frame)

}
