% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hayman.R
\name{HaymanR}
\alias{HaymanR}
\title{Girffing, Hayman ve diğer ıslah desenlerinin analizi}
\usage{
HaymanR(data)
}
\arguments{
\item{data}{A \code{data.frame}}

\item{Rep}{Replicate as factor}

\item{Cross1}{Cross 1 as factor}

\item{Cross2}{Cross 2 as factor}
}
\description{
Bu paket ıslah denemelerinin analizi için geliştirilmiştir.
}
\examples{

# Example1

Hayman(data)



}
