% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LxT.R
\name{LinexTest}
\alias{LinexTest}
\title{Sıra x Dizi ıslah desenlerinin analizi}
\usage{
LinexTest(data)
}
\arguments{
\item{data}{A \code{data.frame}}

\item{replications}{Replications as factor}

\item{lines}{lines  as factor}

\item{testers}{testers  as factor}

\item{y}{y  as variable}
}
\description{
Bu paket ıslah denemelerinin analizi için geliştirilmiştir.
}
\examples{

# Example1

LinexTest(data)


}
