% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NC1.R
\name{NCM1}
\alias{NCM1}
\title{Girffing, Hayman ve diğer ıslah desenlerinin analizi}
\usage{
NCM1(data)
}
\arguments{
\item{data}{A \code{data.frame}}

\item{set}{set as factor}

\item{male}{male 1 as factor}

\item{female}{male 2 as factor}

\item{progenie}{progenie as factor}

\item{replication}{replication as factor}

\item{model}{Model for Diallel Analysis using NC approach.
It can take \strong{1} or \strong{2} or \strong{3} as arguments depending on the model being used.
\enumerate{
\item NCM1 Model;
\item NCM2 Model;
\item NCM3 Model.
}}
}
\description{
Bu paket ıslah denemelerinin analizi için geliştirilmiştir.
}
\examples{

# Example1

NCM1(data)



}
