% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HaymanPlot.R
\name{HaymanPlot}
\alias{HaymanPlot}
\title{Girffing, Hayman ve diğer ıslah desenlerinin analizi}
\usage{
HaymanPlot(data.frame)
}
\arguments{
\item{VOLO}{<- HaymanR$VOLO}

\item{In.Value}{<- HaymanR$In.Value}

\item{a}{<- HaymanR$a}

\item{b}{<- HaymanR$b}

\item{Wr.Vr}{<- HaymanR$Wr.Vr.Table}

\item{data}{A \code{data.frame}}
}
\description{
Bu paket ıslah denemelerinin analizi için geliştirilmiştir.
}
\examples{

# Example1

HaymanPlot(data.frame)

}
